UI and entities update
Updated the UI, now it uses mustache.js to render html.
The entities were also updated.
The podcast list is now loaded at when the background page
is loaded.
Podecast information is still not persisted, only the URL
list is persisted in chrome sync storage.
All podcast data is get from the feeds every time the
bg page is loaded.