Switch from Mustach to AngularJS - part 1
Switched all the views to AngularJS.
The feed finder popup is still using Mustache.
The performance of the rendering went noticeably down, specially when
rendering episode lists with a lot of episodes.

We still need to implement a mechanism to react to changes on the podcasts
and update the controllers.