ngInfiniteScroll for last episodes
This commit adds ngInfiniteScroll
(https://sroze.github.io/ngInfiniteScroll/) to the last episodes
page. Now episodes are loaded as the user scrolls down.

For this to work, we had to include <!DOCTYPE html> in the main
page, because ngInfiniteScroll uses $window.height(), which
does not work as expected without this explicit doctype:
http://stackoverflow.com/questions/12103208/
jquery-window-height-is-returning-the-document-height

As a side effect, the margin for the h1 elements had to be
adjusted and the font sizes have changed.