Performance tweak for update all at last eps view
When update all is triggered, the podcastChanged message is
issued immediately for all podcasts due to the status change.
This created a moment where the episodes list had to be
recreated many times in a row.

Now, the podcast list update is only triggered after the
episode list for a podcast is really updated.

This creates a performance improvement.