First version options page and message service
Created first version of the options page.
Now the options page is there mainly for auto update.

Due to na incosistent behavior of the message API:
https://bugs.chromium.org/p/chromium/issues/detail?id=592478

I had to implement a wrapper for the message API and ensure
that the messages are always sent to background page, even
if no frontground page is open.