Added accepts to ajax setup for loading podcast feeds
This aparently have no effect, but seems more acurate then before.
It overwrites the xml data type to also use application/rss+xml.
It was introduced in an atempt to fix a issue that for feed from
the domain www.deviante.com.br are failing with code
net::ERR_CONTENT_DECODING_FAILED