[player] Fixed default options issue
Default values for the options were handled by users of the
options individually insteady of centrally in the player.

Now they are handled in the player.

It fixes a bug that the default value for "remove from playlist
when finished" was shown as true but it was handled by the player
as false.