[playlist] Fixed issue when episode was removed from feed
When an episode was removed from feed (guid could not be found),
the playlist showed an empty entry.

If more then one episode was removed, then no entry is shown as
there would be duplicates (undefined) entries on the list for
the ng-repeat.

One possible side effect is that when you add a podcast in one
PC, add an episode to the playlist and then go to a different
PC, this entry will be removed in the second PC, if it did not
synch the new podcast yet.

This will probably be fixed when issue #23:
https://bitbucket.org/dellagustin/podstation_chrome_ext/issues/23
is fixed.