Big refactoring: automated tests, feed without guid
This huge commit is a refactoring to accommodate feeds with episodes
without guid.

Automated tests were added, which also required refactorings.

All the commands related to episodes now rely on an episodeId, which
is a unique way to identify a podcast in memory.

What is saved in the sync storage to identify an episode is an
EpisodeSelector, and the selection has the following priority:
* guid
* title

Other fields can be added in the future and the priority below
guid may change without disruption.