delay time to initial updated
After the a big refactoring the initial update o podcasts broke.
It broke because the loading of podcasts from local storage was
deferred, so there was no podcasts to update.

The ideal solution would be to hook up the initial update to
some kind of event that is triggered after the podcasts are all
loaded, but for the moment I simply added a 1 second delay to the
intial podcast update.