Fix jsmediatags dist issue
This commit fixes #47

https://github.com/podStation/podStation/issues/47

Our bower dependency for jsmediatags is "^3.9.0", which means
it picks newer versions as well (I don't know the specifics
by heart, but at least now it is selecting 3.9.3).

From 3.9.0 to 3.9.3 there was a non-backwards compatible change
in distribution (now the minified version is distributed instead),
with a change on the file name.

This was the root cause for the build failure.
I could pin the dependency to 3.9.3 exactly (matter of fact, I
only tought about it while writing this commit), but I see that
the risk of happending again is rather low, so I decided not to
do that.

Reference:
https://github.com/aadsm/jsmediatags/commit/aab8bb0951f3b7c95aeb4a8e37b3f7e59b2c6653#diff-0a08a7565aba4405282251491979bb6b