FFE: Removed illegal characted in locales file
When trying to load the extension in firefox, it complained about
illegal chars in one of hte locales file.
There were tabs (not escaped) at the value of a string in the json
file.
Chrome accepted this, but FF did not.
I remove the chars and then FF could load the extension.