FFE: Added explicit extension id
After loading the extension to firefox, I inspected the background
page's console for errors.

One of the errors was:
 Unchecked lastError value: Error: The storage API will not work with a
 temporary addon ID. Please add an explicit addon ID to your manifest.
 For more information see https://bugzil.la/1323228.

Therefore I generated a Version 4 UUID using the website
https://www.uuidgenerator.net/

Acording to
https://developer.mozilla.org/en-US/docs/Mozilla/Add-ons/WebExtensions/manifest.json/browser_specific_settings
the "gecko" "id" could be either a GUID or a string formatted like a an
email address.

Since podStation does not have its own domain, I decided to use a GUID.