Added search from podcastindex.org
I refactored the search a bit to accomodate the new index
podcastindex.org.

As a part of the refactoring I also itegrated back the code
of https://github.com/podStation/podStationNGReuse into this
repository, as it does not make send to reuse it anymore
from an external source. It was once detatched so it could
be reused, but that does not make much sense today.