Partial imp #165: Support for the tag managingEditor
Partial implementation for
https://github.com/podStation/podStation/issues/165

I did an analysis of around 50000 feeds, I could find only a handful of
feeds that support webMaster.
Non of these feeds used mailto: links on the managingEditor or webMaster
tags, so I will not support this.
I also found out that feeds generated with blogger.com set the email
"noreply@blogger.com" on the managingEditor tag, so I will add special
treatment for that.

I also will probably not implement webMaster now, as this will require
some changes in the UI (like showing it as a support mail), it will be a
left over for another time.