Fix #204 - Removed ajax cachebusting, added HTTP conditional requests
This commit fixes https://github.com/podStation/podStation/issues/204
by removing the ajax cache: false option, which introduced the `_`
query parameter with a timestamp for cachebusting.
This parameter was not accepted by podbean feeds.

I have replaced it by the header Cache-Control.

At the same time, I implemented support for HTTP conditional requests.
I was expecting this to be supported by the browser directly,
but it seems that for ajax requests, it must be done in code.

Ref:
https://developer.mozilla.org/en-US/docs/Web/HTTP/Conditional_requests