v4v: LND Custom Record enablement
With this commit, a single custom record can now be added to
transactions sent with LND, along with the keysend preimage.

This was implemented to enable podStation to get a 1% commission.
In the future, it will also be used by feeds that have a custom
record.
See https://github.com/Podcastindex-org/podcast-namespace/issues/172