Enhancement #212: Avoid jumping player
Changed the font-family of the text that shows the
player time (lenght and current position) to `sans-serif`.

According to my tests on both macOS and Windows (both on Chrome),
it solved the issue of the "jumping player".

The jumping on the player is caused by the width of the current
time changing as the playback progresses.

Also according to my tests, this issue was only observed on mac.

I also tested using FontAwsome fixed width icons on the player
since the transition of the play icon to pause icon and vice
versa also makes the player "jump". The result was not
aestetically pleasing, so I left this out and accepted the
"jumping" on this case, as it happens infrequently.

Reference:
https://github.com/podStation/podStation/issues/212