Refactoring: Bundling with Webpack, automated testing ajax fix
This commit fixes an error on the test helper for ajax.
Reference
https://github.com/podStation/podStation/pull/268#issuecomment-869013529

A XMLHttpRequest object was complaining about the test helper
setting its status.
I don't understand why this error came up now, it was not a problem
so far.
I don't remember why I wrote it this way. In the fix I simply reuse
the actual request object used to request the file content used
for mocking the requests.