Refactoring: Bundling with Webpack, fix automated tests
After the refactoring, some tests were failing.

After investigating, the issue was related to missing message handlers
on the notification manager:
https://github.com/podStation/podStation/blob/70b898ad357928afaeee071c3048accb5ee0ef24/extension/background/entities/notificationManager.js#L73-L78

Mystery solved. Before the refactoring we were including jQuery, now we
don't. Angular will use jQuery if it is there, if it is not, it will use
an embedded component called JQLite. JQLite ready executes asynchronous
even if the document is already loaded, while jQuery executes
synchronously. The test does not work when ready is async.

The reason why we need to add the message listeners asynchronously is a
bit shady, but it is related with the fact that service may be used by
non-angular components before angular bootstrap is done (it was
introduced with
https://github.com/podStation/podStation/commit/5569eacc3aa79d0ca39fcb0f0e49e7b1fa2257ea,
this may shed some light on it).

After including jquery on the tests it worked like a charm.