eslint: minor comparison issue
This does not really fix a bug, since the only message that
was received had message.action = 'getFeeds'.

I tested with breakpoint at the line, I was expecting it
to get messages sent from the pages and background as well.
Somehow, it did not stop there. I cannot really understand
why...

Anyhow, I tested it and it is still working.