Fix #283 - v4v - Wrong encoding of custom value
This commit fixes issue
https://github.com/podStation/podStation/issues/283

podStation was assuming that the attribute `customValue`
of the RSS feed tag `podcast:valueRecipient` was always
a string containing hex enconded binary data.

The spec for this attribute is not specific about its
content, but investigation of current uses (documented
in the issue) showed that it is used to hold strings.

Investigation also showed that some recipients assume
that this strings will be encoded in UTF-8 (1).

While we cannot be sure on which encoding is expected
for every receiver at the moment, the known examples
show that the charaters in use are part of the 7bit
range of ASCII (not extented ASCII).

Withing this range, UTF-8 is compatible with many other
encodings, and thus its usage should be acceptable, until
we have a more strict definition on how this field should
be treated.

I have sent a Pull Request proposing that the content of
this field must be treated as a string and always encoded
with UTF-8 when added to the TLV stream of lightning payments.
PR: https://github.com/Podcastindex-org/podcast-namespace/pull/307

1:
https://github.com/Podcastindex-org/podcast-namespace/pull/307#issuecomment-944960358