v4v: Fix #289 Caching LNPay Balance
This commit introduces caching for LNPay balance, addressing
https://github.com/podStation/podStation/issues/289.

A few days ago my IP was banned from LNPay due to excessive
calls to the get the wallet balance.

After cheching the code, it was really sending to many
requests.

This commit removes some unecessary calls to get the
balance and also stablishes a 60 second cache for
the balance.

Additionally, the response from LNPay payments
contains the wallet balance, so I am also caching
the balance of the responses.

With this, I don't expect podStation to send
balance requests anymore during playback.