Fix #325 - Angular 1.8.3 - Fixed navigation
Since Angular 1.6 the default hash prefix is '!' instead of ''.

It would be possible to resolve this with the following code
according to the changelog of AngularJS:
```
appModule.config(['$locationProvider', function($locationProvider) {
  $locationProvider.hashPrefix("");
  }]);
```

Since the default changed, I decided to adopt the defautl instead
and adjusted the code.Since the default changed, I decided to adopt the
defautl instead
and adjusted the code.

References:
- https://stackoverflow.com/questions/44220693/exclamation-in-routes-using-angular-ui-router
- https://stackoverflow.com/questions/41734185/hashbang-defaulting-in-angular-application
- https://github.com/angular/angular.js/blob/master/CHANGELOG.md#location-due-to
- https://github.com/angular/angular.js/commit/aa077e81129c740041438688dff2e8d20c3d7b52

Fixes:
- https://github.com/podStation/podStation/issues/325